Refactor: Dto packaging
