+++ src/main/java/com/capstone/smutaxi/controller/MatchingController.java
@@ -1,13 +1,10 @@
 package com.capstone.smutaxi.controller;
 
-import com.capstone.smutaxi.dto.requests.MatchingRequest;
-import com.capstone.smutaxi.dto.responses.MatchingResponseDto;
-import com.capstone.smutaxi.entity.User;
+import com.capstone.smutaxi.dto.requests.match.MatchingRequest;
+import com.capstone.smutaxi.dto.responses.match.MatchingResponseDto;
 import com.capstone.smutaxi.repository.UserRepository;
 import com.capstone.smutaxi.service.matching.MatchingService;
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
