+++ src/main/java/com/capstone/smutaxi/dto/responses/auth/EmailVerificationResponse.java
@@ -1,14 +1,13 @@
-package com.capstone.smutaxi.dto.responses;
+package com.capstone.smutaxi.dto.responses.auth;
 
-import com.capstone.smutaxi.dto.UserDto;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter @Setter
 @Builder
-public class UserUpdateResponse {
+public class EmailVerificationResponse {
     private Boolean success;
     private String message;
-    private UserDto userDto;
+    private Integer verificationCode;
 }
