+++ src/main/java/com/capstone/smutaxi/dto/responses/user/UploadImageResponse.java
@@ -1,0 +1,14 @@
+package com.capstone.smutaxi.dto.responses.user;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Builder
+@Getter @Setter
+public class UploadImageResponse {
+    private Boolean success;
+    private String message;
+    private String imageUrl;
+
+}
