+++ src/main/java/com/capstone/smutaxi/dto/responses/user/UserUpdateResponse.java
@@ -1,13 +1,14 @@
-package com.capstone.smutaxi.dto.responses;
+package com.capstone.smutaxi.dto.responses.user;
 
+import com.capstone.smutaxi.dto.UserDto;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter @Setter
 @Builder
-public class EmailVerificationResponse {
+public class UserUpdateResponse {
     private Boolean success;
     private String message;
-    private Integer verificationCode;
+    private UserDto userDto;
 }
